# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TEMPO ADHD-friendly AI daily planner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /folders:
    get:
      operationId: listFolders
      tags: [folders]
      summary: List all folders
      responses:
        "200":
          description: List of folders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Folder"
    post:
      operationId: createFolder
      tags: [folders]
      summary: Create a new folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFolderBody"
      responses:
        "201":
          description: Created folder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Folder"
  /folders/{id}:
    patch:
      operationId: updateFolder
      tags: [folders]
      summary: Update a folder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFolderBody"
      responses:
        "200":
          description: Updated folder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Folder"
        "404":
          description: Not found
    delete:
      operationId: deleteFolder
      tags: [folders]
      summary: Delete a folder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /daily-plans:
    get:
      operationId: listDailyPlans
      tags: [dailyPlans]
      summary: List daily plans
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of daily plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyPlan"
    post:
      operationId: createDailyPlan
      tags: [dailyPlans]
      summary: Create or update a daily plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDailyPlanBody"
      responses:
        "201":
          description: Created daily plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyPlan"
  /daily-plans/{id}:
    get:
      operationId: getDailyPlan
      tags: [dailyPlans]
      summary: Get a daily plan by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Daily plan details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyPlan"
        "404":
          description: Not found
    patch:
      operationId: updateDailyPlan
      tags: [dailyPlans]
      summary: Update a daily plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDailyPlanBody"
      responses:
        "200":
          description: Updated daily plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyPlan"
        "404":
          description: Not found
...
  /memories:
    get:
      operationId: listMemories
      tags: [preferences]
      summary: List memory items
      responses:
        "200":
          description: List of memory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemoryItem"
    post:
      operationId: createMemory
      tags: [preferences]
      summary: Create a memory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMemoryBody"
      responses:
        "201":
          description: Created memory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryItem"
  /memories/{id}:
    delete:
      operationId: deleteMemory
      tags: [preferences]
      summary: Delete a memory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai]
      summary: Send a message to the AI assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatBody"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatResponse"
  /ai/extract-tasks:
    post:
      operationId: aiExtractTasks
      tags: [ai]
      summary: Extract tasks from messy text input
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiExtractBody"
      responses:
        "200":
          description: Extracted tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiExtractResponse"
  /ai/chunk-task:
    post:
      operationId: aiChunkTask
      tags: [ai]
      summary: Break a large task into smaller subtasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChunkBody"
      responses:
        "200":
          description: Chunked subtasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChunkResponse"
  /ai/prioritize:
    post:
      operationId: aiPrioritize
      tags: [ai]
      summary: Suggest priority ordering for tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiPrioritizeBody"
      responses:
        "200":
          description: Prioritized tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiPrioritizeResponse"
  /ai/generate-plan:
    post:
      operationId: aiGeneratePlan
      tags: [ai]
      summary: Generate a daily plan from tasks and preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiGeneratePlanBody"
      responses:
        "200":
          description: Generated daily plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiGeneratePlanResponse"
...
components:
  schemas:
...
    Folder:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        parentFolderId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdAt
    CreateFolderBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        parentFolderId:
          type: integer
          nullable: true
      required:
        - name
    UpdateFolderBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        parentFolderId:
          type: integer
          nullable: true
...
    DailyPlan:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
        blocks:
          type: array
          items:
            type: object
            additionalProperties: true
        mood:
          type: string
          nullable: true
        energyLevel:
          type: integer
          nullable: true
        aiGenerated:
          type: boolean
        acceptedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - date
        - blocks
        - aiGenerated
        - createdAt
        - updatedAt
    CreateDailyPlanBody:
      type: object
      properties:
        date:
          type: string
        blocks:
          type: array
          items:
            type: object
            additionalProperties: true
        mood:
          type: string
          nullable: true
        energyLevel:
          type: integer
          nullable: true
        aiGenerated:
          type: boolean
          default: false
      required:
        - date
        - blocks
    UpdateDailyPlanBody:
      type: object
      properties:
        blocks:
          type: array
          items:
            type: object
            additionalProperties: true
        mood:
          type: string
          nullable: true
        energyLevel:
          type: integer
          nullable: true
        acceptedAt:
          type: string
          format: date-time
          nullable: true
...
    MemoryItem:
      type: object
      properties:
        id:
          type: integer
        tier:
          type: string
          enum: [hot, warm, cold]
        content:
          type: string
        decay:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - tier
        - content
        - decay
        - createdAt
    CreateMemoryBody:
      type: object
      properties:
        tier:
          type: string
          enum: [hot, warm, cold]
        content:
          type: string
        decay:
          type: integer
          default: 100
      required:
        - tier
        - content
    AiChatBody:
      type: object
      properties:
        message:
          type: string
        context:
          type: object
          additionalProperties: true
          nullable: true
      required:
        - message
    AiChatResponse:
      type: object
      properties:
        response:
          type: string
        suggestions:
          type: array
          items:
            type: string
      required:
        - response
        - suggestions
    AiExtractBody:
      type: object
      properties:
        text:
          type: string
      required:
        - text
    AiExtractResponse:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/ExtractedTask"
      required:
        - tasks
    ExtractedTask:
      type: object
      properties:
        title:
          type: string
        priority:
          type: string
          enum: [high, medium, low]
        estimatedMinutes:
          type: integer
          nullable: true
        tags:
          type: array
          items:
            type: string
      required:
        - title
        - priority
    AiChunkBody:
      type: object
      properties:
        taskTitle:
          type: string
        taskNotes:
          type: string
          nullable: true
      required:
        - taskTitle
    AiChunkResponse:
      type: object
      properties:
        subtasks:
          type: array
          items:
            $ref: "#/components/schemas/ExtractedTask"
        reasoning:
          type: string
      required:
        - subtasks
        - reasoning
    AiPrioritizeBody:
      type: object
      properties:
        tasks:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              title:
                type: string
              priority:
                type: string
              estimatedMinutes:
                type: integer
                nullable: true
            required:
              - id
              - title
      required:
        - tasks
    AiPrioritizeResponse:
      type: object
      properties:
        orderedTaskIds:
          type: array
          items:
            type: integer
        reasoning:
          type: string
      required:
        - orderedTaskIds
        - reasoning
    AiGeneratePlanBody:
      type: object
      properties:
        date:
          type: string
        taskIds:
          type: array
          items:
            type: integer
      required:
        - date
    AiGeneratePlanResponse:
      type: object
      properties:
        blocks:
          type: array
          items:
            type: object
            additionalProperties: true
        reasoning:
          type: string
      required:
        - blocks
        - reasoning
...