# lib/api-zod/src/generated/types/aiChunkResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TEMPO ADHD-friendly AI daily planner API
 * OpenAPI spec version: 0.1.0
 */
import type { ExtractedTask } from "./extractedTask";

export interface AiChunkResponse {
  subtasks: ExtractedTask[];
  reasoning: string;
}