# lib/api-zod/src/generated/types/createFolderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TEMPO ADHD-friendly AI daily planner API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateFolderBody {
  name: string;
  description?: string | null;
  parentFolderId?: number | null;
}