# lib/api-zod/src/generated/types/dailyPlan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TEMPO ADHD-friendly AI daily planner API
 * OpenAPI spec version: 0.1.0
 */
import type { DailyPlanBlocksItem } from "./dailyPlanBlocksItem";

export interface DailyPlan {
  id: number;
  date: string;
  blocks: DailyPlanBlocksItem[];
  mood?: string | null;
  energyLevel?: number | null;
  aiGenerated: boolean;
  acceptedAt?: Date | null;
  createdAt: Date;
  updatedAt: Date;
}