# lib/api-zod/src/generated/types/extractedTask.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TEMPO ADHD-friendly AI daily planner API
 * OpenAPI spec version: 0.1.0
 */
import type { ExtractedTaskPriority } from "./extractedTaskPriority";

export interface ExtractedTask {
  title: string;
  priority: ExtractedTaskPriority;
  estimatedMinutes?: number | null;
  tags?: string[];
}