# lib/api-zod/src/generated/types/updateDailyPlanBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TEMPO ADHD-friendly AI daily planner API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateDailyPlanBodyBlocksItem } from "./updateDailyPlanBodyBlocksItem";

export interface UpdateDailyPlanBody {
  blocks?: UpdateDailyPlanBodyBlocksItem[];
  mood?: string | null;
  energyLevel?: number | null;
  acceptedAt?: Date | null;
}